Добавлена страница "Моё обучение" на стороне обучающегося, которая отображает список заявок и страницу-заглушку списка активных курсов

+++ src/main/java/dev/lms/controllers/RequestController.java
@@ -88,7 +88,7 @@ public class RequestController {
             }
 
             // Получаем заявки студента
-            List<RequestDTO> requests = requestService.getRequestsByStudent(studentId.longValue());
+            List<RequestDTO> requests = requestService.getRequestsByStudent(studentId);
             return ResponseEntity.ok(requests);
 
         } catch (Exception e) {
